[BAU] rename the towns fund round 1 transformation module to specify the fund
